replace all numeric controls

+++ GUI/Controls/SettingsControl.cs
@@ -26,7 +26,7 @@ namespace GUI.Forms
 
             maxTextureSizeInput.Value = Settings.Config.MaxTextureSize;
             shadowResolutionInput.Value = Settings.Config.ShadowResolution;
-            fovInput.Value = (decimal)Settings.Config.FieldOfView;
+            fovInput.Value = Settings.Config.FieldOfView;
             vsyncCheckBox.Checked = Settings.Config.Vsync != 0;
             displayFpsCheckBox.Checked = Settings.Config.DisplayFps != 0;
             openExplorerOnStartCheckbox.Checked = Settings.Config.OpenExplorerOnStart != 0;
@@ -131,7 +131,7 @@ namespace GUI.Forms
                 return;
             }
 
-            Settings.Config.MaxTextureSize = (int)maxTextureSizeInput.Value;
+            Settings.Config.MaxTextureSize = maxTextureSizeInput.Value;
         }
 
         private void OnFovValueChanged(object sender, EventArgs e)
@@ -147,7 +147,7 @@ namespace GUI.Forms
         private void OnSetFovTo4by3ButtonClick(object sender, EventArgs e)
         {
             Settings.Config.FieldOfView = 2f * MathF.Atan(3f / 4f) / MathF.PI * 180f;
-            fovInput.Value = (decimal)Settings.Config.FieldOfView;
+            fovInput.Value = Settings.Config.FieldOfView;
         }
 
         private void OnOpenExplorerOnStartValueChanged(object sender, EventArgs e)
